ui/gtk-egl: fixes compiling error
tex_id is uint32_t but force convert to EGLClientBuffer and that's a pointer

+++ ui/egl-helpers.c
@@ -225,7 +225,7 @@ int egl_get_fd_for_texture(uint32_t tex_id, EGLint *stride, EGLint *fourcc,
 
     image = eglCreateImageKHR(qemu_egl_display, eglGetCurrentContext(),
                               EGL_GL_TEXTURE_2D_KHR,
-                              (EGLClientBuffer)(unsigned long)tex_id,
+                              (EGLClientBuffer)(uintptr_t)tex_id,
                               NULL);
     if (!image) {
         return -1;
